+++ libdevcore/Base64.cpp
