+++ libdevcore/Common.cpp
