+++ libdevcore/CommonData.cpp
