+++ libdevcore/CommonIO.cpp
