+++ libdevcore/CommonIO.h
