+++ libdevcore/CommonJS.cpp
