+++ libdevcore/CommonJS.h
