+++ libdevcore/Exceptions.h
