+++ libdevcore/Guards.cpp
