+++ libdevcore/Guards.h
