+++ libdevcore/Log.cpp
