+++ libdevcore/Log.h
