+++ libdevcore/MemoryDB.cpp
