+++ libdevcore/MemoryDB.h
