+++ libdevcore/Readme.md
@@ -1,0 +1,17 @@
+# relational db
+
+using cpp-ethereum merkle patricia tree
+
+## compile
+
+`chmod +x build_server_client.sh`
+
+`./build_server_client`
+
+## run
+
+`./server`
+
+on multiple terminals
+
+`./client`
