+++ libdevcore/Terminal.h
