+++ libdevcore/TrieCommon.cpp
