+++ libdevcore/TrieCommon.h
