+++ libdevcore/TrieHash.cpp
