+++ libdevcore/TrieHash.h
