+++ libdevcore/Worker.cpp
