+++ libdevcore/Worker.h
