+++ libdevcore/concurrent_queue.h
