+++ libdevcore/vector_ref.h
